purchase controller

+++ Ibdal.Models/Product.cs
@@ -6,7 +6,7 @@ public class Product : Archivable
 {
     public required string Name { get; set; }
     public required string Category { get; set; }
-    public required decimal Price { get; set; }
+    public required int Price { get; set; }
     public required string Description { get; set; }
     public required string ImageUrl { get; set; }
 }
